lots of errors

+++ src/main/java/com/robertx22/mine_and_slash/database/spells/spell_classes/bases/level_based_numbers/LevelBased.java
@@ -49,7 +49,7 @@ public class LevelBased {
             return maxLevel;
         }
 
-        float multi = abilityLevel / ability.getMaxSpellLevelBuffed();
+        float multi = abilityLevel / (float) ability.getMaxSpellLevelBuffed();
 
         float val = levelOne + ((maxLevel - levelOne) * multi);
